requirements.txt
pyocr
Pillow==8.0.1
opencv-python